requirements.txt
web3==6.10.0
flask==3.0.2
aiohttp==3.10.11
requests==2.32.0
python-dotenv==1.0.1
pytest>=7.0.0
requests>=2.0.0
setuptools==65.0.0